+++ lib/kelp/checkbox.rb
@@ -7,6 +7,18 @@ module Kelp
     include Scoping
     include Helper
 
+    # Verify that the given checkbox is checked.
+    #
+    # @param [String] checkbox
+    #   Capybara locator for the checkbox
+    # @param [Hash] scope
+    #   Scoping keywords as understood by {#in_scope}
+    #
+    # @raise [Kelp::Unexpected]
+    #   If the given checkbox is not checked
+    #
+    # @since 0.1.2
+    #
     def checkbox_should_be_checked(checkbox, scope={})
       in_scope(scope) do
         field_checked = find_field(checkbox)['checked']
@@ -18,6 +30,19 @@ module Kelp
       end
     end
 
+
+    # Verify that the given checkbox is not checked.
+    #
+    # @param [String] checkbox
+    #   Capybara locator for the checkbox
+    # @param [Hash] scope
+    #   Scoping keywords as understood by {#in_scope}
+    #
+    # @raise [Kelp::Unexpected]
+    #   If the given checkbox is checked
+    #
+    # @since 0.1.2
+    #
     def checkbox_should_not_be_checked(checkbox, scope={})
       in_scope(scope) do
         field_checked = find_field(checkbox)['checked']