+++ spec/attribute_spec.rb
@@ -15,13 +15,13 @@ describe Kelp::Attribute, "should_be_disabled" do
     it "element does not exist" do
       lambda do
         should_be_disabled "nonexistent"
-      end.should raise_error(rspec_unexpected)
+      end.should raise_error(Kelp::Unexpected)
     end
 
     it "element does not have the disabled attribute" do
       lambda do
         should_be_disabled "first_name"
-      end.should raise_error(rspec_unexpected)
+      end.should raise_error(Kelp::Unexpected)
     end
   end
 end
@@ -42,13 +42,13 @@ describe Kelp::Attribute, "should_be_enabled" do
     it "element does not exist" do
       lambda do
         should_be_enabled "nonexistent"
-      end.should raise_error(rspec_unexpected)
+      end.should raise_error(Kelp::Unexpected)
     end
 
     it "element has the disabled attribute" do
       lambda do
         should_be_enabled "readonly"
-      end.should raise_error(rspec_unexpected)
+      end.should raise_error(Kelp::Unexpected)
     end
   end
 end