+++ spec/checkbox_spec.rb
@@ -20,14 +20,14 @@ describe Kelp::Checkbox, "checkbox_should_be_checked" do
     it "the checkbox is unchecked by default" do
       lambda do
         checkbox_should_be_checked "I like salami"
-      end.should raise_error
+      end.should raise_error(Kelp::Unexpected)
     end
 
     it "the checkbox is unchecked programmatically" do
       uncheck "I like cheese"
       lambda do
         checkbox_should_be_checked "I like cheese"
-      end.should raise_error
+      end.should raise_error(Kelp::Unexpected)
     end
   end
 
@@ -54,14 +54,14 @@ describe Kelp::Checkbox, "checkbox_should_not_be_checked" do
     it "the checkbox is checked by default" do
       lambda do
         checkbox_should_not_be_checked "I like cheese"
-      end.should raise_error
+      end.should raise_error(Kelp::Unexpected)
     end
 
     it "the checkbox is checked programmatically" do
       check "I like salami"
       lambda do
         checkbox_should_not_be_checked "I like salami"
-      end.should raise_error
+      end.should raise_error(Kelp::Unexpected)
     end
   end